VEC-598: write run_id on identity tables, bind a zero RunID as NULL
Review follow-ups: the 13 governed identity/config tables (token, user, protocol,
receipt_token, debt_token, fluid_vault, maple_*, morpho_market, morpho_vault, oracle,
protocol_oracle) now carry run_id on insert, so their new rows are no longer
indistinguishable from pre-tracking rows; RunID is a driver.Valuer that binds NULL for
zero instead of naming a run that does not exist; OpenRun moves to run.go beside it.
Test setup deduped: SetDevIdentity carries the stand-in git hash, RequireRunID replaces
the copied read-back assertion.

+++ stl-verify/internal/adapters/outbound/postgres/allocation_repository.go
@@ -203,7 +203,7 @@ func (r *AllocationRepository) buildInsertArgs(
 		pos.Direction,
 		pos.CreatedAt,
 		int(r.buildID),
-		int64(r.runID),
+		r.runID,
 		underlyingValue,
 		underlyingTokenID,
 		encodeAddress(pos.FromAddress),
