+++ stl-verify/internal/adapters/outbound/postgres/prime_balance_sheet_repository.go
@@ -89,7 +89,7 @@ func (r *PrimeBalanceSheetRepository) SaveBalanceSheetSnapshots(
 				s.BackstopCapitalUSD,
 				s.Source,
 				s.BuildID,
-				int64(r.runID),
+				r.runID,
 			)
 		}
