+++ stl-verify/internal/adapters/outbound/postgres/prime_capital_stack_allocation_repository.go
@@ -97,7 +97,7 @@ func (r *PrimeCapitalStackAllocationRepository) SaveCapitalStackAllocations(
 			a.CRR,
 			a.Source,
 			a.BuildID,
-			int64(r.runID),
+			r.runID,
 		)
 	}
